requirements.txt
PyQt5==5.15.11
openai==2.3.0
gTTS==2.5.4
pyttsx3==2.90
SpeechRecognition==3.14.3
pyaudio==0.2.14
pyautogui==0.9.54
pytesseract==0.3.13
Pillow==12.0.0
typing-extensions==4.15.0
tqdm==4.67.1
httpx==0.28.1
requests==2.32.5
pygetwindow==0.0.9
pyscreeze==1.0.1
pytweening==1.2.0
